refactor: renamed period utils
