+++ src/components/PeriodDimension/utils/fixedPeriods.js
