+++ src/components/PeriodDimension/utils/relativePeriods.js
